refactor(Scores): makeabstract class
다형성 적용하여 score를 더하는 부분의 분기문을 제거.
