+++ src/main/java/bowling/domain/scores/Scores.java
@@ -1,5 +1,7 @@
-package bowling.domain.score;
+package bowling.domain.scores;
 
+import bowling.domain.score.Score;
+import bowling.domain.score.ScoreType;
 import org.springframework.util.CollectionUtils;
 
 import java.util.ArrayList;
@@ -8,18 +10,27 @@ import java.util.List;
 /**
  * 각 프레임에서 기록한 점수들를 저장한다.
  */
-public class Scores {
+public abstract class Scores {
+    private static final int ONE = 1;
     private static final int SECOND_PLAY = 1;
     private static final int THIRD_PLAY = 2;
+    private static final int DEFAULT_PLAY_COUNT = 2;
 
     private final List<Score> scores;
 
-    public Scores() {
+    protected Scores() {
         scores = new ArrayList<>();
     }
 
-    public void add(Score score) {
-        scores.add(score);
+    protected abstract Score createScore(int point);
+
+
+    public void add(int point) {
+        scores.add(createScore(point));
+    }
+
+    protected Score getLastScore() {
+        return scores.get(scores.size() - ONE);
     }
 
     public int size() {
@@ -60,5 +71,9 @@ public class Scores {
 
     public boolean isThirdPlay() {
         return scores.size() == THIRD_PLAY;
+    }
+
+    public boolean isPlayable() {
+        return scores.size() < DEFAULT_PLAY_COUNT;
     }
 }
